+++ src/assets/scss/_mixin.scss
@@ -1,0 +1,5 @@
+@mixin lg {
+  @media (min-width: 992px) {
+    @content;
+  }
+}
