+++ src/assets/scss/main.scss
@@ -1,4 +1,3 @@
-// @import './bootstrap';
 @import './variables';
 @import url('https://fonts.googleapis.com/css2?family=Noto+Serif+TC&display=swap');
 @import url('https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200');
@@ -21,3 +20,11 @@ a{
 .pointer{
   cursor: pointer;
 }
+.title {
+  @extend .fs-6;
+  @extend .fs-lg-5;
+  @extend .fw-bold;
+  @extend .mb-5;
+  @extend .mb-lg-7;
+}
+
