+++ src/components/UserCenter/NextRoom.vue
@@ -1,0 +1,17 @@
+<template>
+    <p class="fw-medium text-neutral-80">預訂參考編號： HH2302183151222</p>
+    <h2 class="title">即將來的行程</h2>
+    <img src="https://github.com/hexschool/2022-web-layout-training/blob/main/typescript-hotel/%E6%A1%8C%E6%A9%9F%E7%89%88/room2-1.png?raw=true" alt="" class="nextRoomImg rounded-3 mb-5 mb-lg-7">
+</template>
+<style lang="scss" scoped>
+
+.nextRoomImg{
+  width: 100%;
+  object-fit: cover;
+  object-position: bottom;
+  height: 150px;
+  @include lg{
+   height: 240px;
+  }
+}
+</style>
