+++ vite.config.ts
@@ -17,7 +17,9 @@ export default defineConfig({
   },
   css: {
     preprocessorOptions: {
-      scss: { additionalData: `@import "@/assets/scss/bootstrap";` },
+      scss: { 
+        additionalData: `@import "@/assets/scss/bootstrap";` 
+      },
     },
   }
 })
